refactor: abstracted out ProductItem from ProductServiceBox
